[REFACTOR] (raceCar) : raceCar 폴더구조 변경
- component 폴더 아래 있던 raceCar를 src 밑으로 수정
(input과 output이 class로 분류되면서 함께 컴포넌트라고 지칭하기엔 모호해짐....)

raceCar의 폴더구조를 변경하였습니다.
